+++ NEWS.md
@@ -12,6 +12,7 @@ from the US EPA, AirNow, AIRSIS, WRCC and others.
 ### PWFSLSmoke 1.0.15
 
  * `wrcc_downloadData()` now uses `parseDatetime()` like other functions
+ * `monitorID` argument is no longer required in `monitor_join()`. Defaults to joining all shared monitorIDs.
  
 ### PWFSLSmoke 1.0.14